requirements.txt
Flask
tensorflow
Pillow
scikit-learn
